% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.plusminus}
\alias{get.plusminus}
\title{Bias-adjusted Gaussian confidence intervals.}
\usage{
get.plusminus(max.bias, sampling.se, alpha = 0.95)
}
\arguments{
\item{max.bias}{Worst-case bias of estimate.}

\item{sampling.se}{Sampling error of estimate.}

\item{alpha}{Coverage probability of confidence interval.}
}
\value{
Half-width of confidence interval.
}
\description{
Bias-adjusted Gaussian confidence intervals.
}
