% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.lin.reg.R
\name{llr}
\alias{llr}
\title{Locally linear regression discontinuity design}
\usage{
llr(X, max.second.derivative, bandwidth = NULL, Y = NULL,
  num.samples = rep(1, length(X)), threshold = 0, sigma.sq = NULL,
  change.derivative = TRUE, alpha = 0.95, max.window = max(abs(X -
  threshold)), num.bucket = 200, kernel = c("rectangular", "triangular"),
  minimization.target = c("mse", "ci.length"),
  use.homoskedatic.variance = FALSE)
}
\arguments{
\item{X}{The running variables.}

\item{max.second.derivative}{A bound on the second derivative of mu_w(x) = E[Y(w) | X = x].}

\item{bandwidth}{Bandwidth for the llr. If null, adaptively optimize for the bandwidth.}

\item{Y}{The outcomes.}

\item{num.samples}{Number of samples used to compute each datapoint (perhaps we only have
access to summary data, where each datapoint is averaged over many samples).}

\item{threshold}{The threshold determining treatment assignment.}

\item{sigma.sq}{The irreducible noise level. If null, estimated from the data.}

\item{change.derivative}{Whether we allow for a change in slope at the threshold.}

\item{alpha}{Coverage probability of confidence intervals.}

\item{max.window}{Maximum possible bandwidth to consider in optimization.}

\item{num.bucket}{Number of histogram buckets in numerical analysis.}

\item{kernel}{Shape of weighting function for local regression.}

\item{minimization.target}{Whether bandwidth should minimize maximum mse or confidence
interval length.}

\item{use.homoskedatic.variance}{Whether confidence intervals should be built assuming homoskedasticity.}
}
\value{
A trained optrdd object. Note that locally linear regression also provides
        a linear estimator for tau, just like optrdd. The gammas simply aren't optimal.
}
\description{
Locally linear estimation and inference of treamtment effects identified
via regression discontinuities
}
