% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd.weights.2d.R
\name{optrdd.2d}
\alias{optrdd.2d}
\title{Optimized regression discontinuity design with bivariate running variable}
\usage{
optrdd.2d(X, max.second.derivative, Y = NULL, threshold = c(0, 0),
  sigma.sq = NULL, estimate.cate.at.point = FALSE,
  center.treated.sample = FALSE, alpha = 0.95, lambda.mult = 1,
  max.window = c(max(abs(X[, 1] - threshold[1])), max(abs(X[, 2] -
  threshold[2]))), num.bucket = c(20, 20))
}
\arguments{
\item{X}{The running variables (must be an n x 2 matrix).}

\item{max.second.derivative}{A bound on the operator norm of the second derivative of
mu_w(x) = E[Y(w) | X = x].}

\item{Y}{The outcomes.}

\item{threshold}{A vector (t1, t2) such that samples are treated if Xi[1] < t1 and Xi[2] < t2.}

\item{sigma.sq}{The irreducible noise level. If null, estimated from the data.}

\item{estimate.cate.at.point}{Whether we want to estimate the CATE at the threshold, or can
allow for some weighted average of the CATE.}

\item{center.treated.sample}{Whether the gamma-weighted average of the treated sample should
match the threshold. Note: if estimate.cate.at.point is TRUE, the
center.treated.sample should also be TRUE.}

\item{alpha}{Coverage probability of confidence intervals.}

\item{lambda.mult}{Optional multplier that can be used to over- or under-penalize variance.}

\item{max.window}{Observations further than max.window from the threshold are ignored.
This parameter is a length-2 vector; both coordinates mus wall within the window
to be used.}

\item{num.bucket}{Number of buckets used in numerical optimization.}
}
\value{
Trained optrdd.2d object.
}
\description{
Optimized estimation and inference of treamtment effects identified
via bivariate regression discontinuities
}
