% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optrdd.R
\name{optrdd}
\alias{optrdd}
\title{Optimized regression discontinuity design}
\usage{
optrdd(X, Y = NULL, W, max.second.derivative, estimation.point = NULL,
  sigma.sq = NULL, alpha = 0.95, lambda.mult = 1, bin.width = NULL,
  use.homoskedatic.variance = FALSE, use.spline = TRUE, spline.df = NULL,
  optimizer = c("auto", "mosek", "quadprog"), verbose = TRUE)
}
\arguments{
\item{X}{The running variables.}

\item{Y}{The outcomes. If null, only optimal weights are computed.}

\item{W}{Treatment assignments, typically of the form 1(X >= c).}

\item{max.second.derivative}{A bound on the second derivative of mu_w(x) = E[Y(w) | X = x].}

\item{estimation.point}{Point "c" at which CATE is to be estimated. If estimation.point = NULL,
we estimate a weighted CATE, with weights chosen to minimize MSE,
as in Section 4.1 of Imbens and Wager (2017).}

\item{sigma.sq}{The irreducible noise level. If null, estimated from the data.}

\item{alpha}{Coverage probability of confidence intervals.}

\item{lambda.mult}{Optional multplier that can be used to over- or under-penalize variance.}

\item{bin.width}{Bin width for discrete approximation.}

\item{use.homoskedatic.variance}{Whether confidence intervals should be built assuming homoskedasticity.}

\item{use.spline}{Whether non-parametric components should be modeled as quadratic splines
in order to reduce the number of optimization parameters, and potentially
improving computational performance.}

\item{spline.df}{Number of degrees of freedom (per running variable) used for spline computation.}

\item{optimizer}{Which optimizer to use? Mosek is a commercial solver, but free
academic licenses are available. Needs to be installed separately.
Quadprog is the default R solver; it may be slow on large problems, but
is very accurate on small problems. The option "auto" uses a heuristic to choose.}

\item{verbose}{whether the optimizer should print progress information}
}
\value{
A trained optrdd object.
}
\description{
Optimized estimation and inference of treamtment effects identified
via regression discontinuities
}
\examples{

}
