% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd.weights.R
\name{optrdd}
\alias{optrdd}
\title{Optimized regression discontinuity design}
\usage{
optrdd(X, max.second.derivative, Y = NULL, num.samples = rep(1, length(X)),
  threshold = 0, sigma.sq = NULL, change.derivative = TRUE,
  alpha = 0.95, lambda.mult = 1, max.window = max(abs(X - threshold)),
  num.bucket = 200, use.homoskedatic.variance = FALSE)
}
\arguments{
\item{X}{The running variables.}

\item{max.second.derivative}{A bound on the second derivative of mu_w(x) = E[Y(w) | X = x].}

\item{Y}{The outcomes.}

\item{num.samples}{Number of samples used to compute each datapoint (perhaps we only have
access to summary data, where each datapoint is averaged over many samples).}

\item{threshold}{The threshold determining treatment assignment.}

\item{sigma.sq}{The irreducible noise level. If null, estimated from the data.}

\item{change.derivative}{Whether we allow for a change in slope at the threshold.}

\item{alpha}{Coverage probability of confidence intervals.}

\item{lambda.mult}{Optional multplier that can be used to over- or under-penalize variance.}

\item{max.window}{Observations further than max.window from the threshold are ignored.}

\item{num.bucket}{Number of buckets used in numerical optimization.}

\item{use.homoskedatic.variance}{Whether confidence intervals should be built assuming homoskedasticity.}
}
\value{
A trained optrdd object.
}
\description{
Optimized estimation and inference of treamtment effects identified
via regression discontinuities
}
